sopnode/ansible/requirements.txt
pynetbox==7.0.0
PyYAML==6.0
requests==2.22.0